+++ spherical_geometry/tests/test_basic.py
@@ -8,7 +8,11 @@ import pytest
 from numpy.testing import assert_almost_equal, assert_allclose
 
 from spherical_geometry import graph, great_circle_arc, polygon, vector
-from spherical_geometry.tests.helpers import ROOT_DIR, get_point_set, resolve_imagename
+from spherical_geometry.tests.helpers import (
+    ROOT_DIR,
+    get_point_set,
+    resolve_imagename
+)
 
 try:
     from spherical_geometry import math_util
@@ -536,3 +540,30 @@ def test_math_util_angle_nearly_coplanar_vec():
 
     assert_allclose(angles[:-1], np.pi, rtol=0, atol=1e-16)
     assert_allclose(angles[-1], 0, rtol=0, atol=1e-32)
+
+
+def test_inner1d():
+    vectors = [
+        [1.0, 1.0, 1.0],
+        3 * [1.0 / np.sqrt(3)],
+        [1, 0.5, 1],
+        [1, 0.15, 1],
+        [-1, 0.1, -1]
+    ]
+    lengths = great_circle_arc._inner1d_np(vectors, vectors)
+
+    assert_allclose(lengths, [3.0, 1.0, 2.25, 2.0225, 2.01], rtol=0, atol=1e-15)
+
+
+@pytest.mark.skipif(math_util is None, reason="math_util C-ext is missing")
+def test_math_util_inner1d():
+    vectors = [
+        [1.0, 1.0, 1.0],
+        3 * [1.0 / np.sqrt(3)],
+        [1, 0.5, 1],
+        [1, 0.15, 1],
+        [-1, 0.1, -1]
+    ]
+    lengths = math_util.inner1d(vectors, vectors)
+
+    assert_allclose(lengths, [3.0, 1.0, 2.25, 2.0225, 2.01], rtol=0, atol=1e-15)